Hateoas refactoring
