+++ build.gradle
@@ -26,19 +26,24 @@ subprojects {
     }
 
     dependencies {
+        implementation group: 'org.springframework.boot', name: 'spring-boot-starter-parent', version: '2.4.3', ext: 'pom'
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+        implementation 'org.springframework.boot:spring-boot-starter-hateoas'
+        implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.boot:spring-boot-starter'
+        implementation group: 'org.springframework.boot', name: 'spring-boot-starter-actuator', version: '2.4.3'
         compile group: 'org.modelmapper', name: 'modelmapper', version: '2.3.9'
         compile group: 'net.jodah', name: 'typetools', version: '0.6.2'
         compile group: 'com.mchange', name: 'c3p0', version: '0.9.5.2'
         compile group: 'org.hibernate', name: 'hibernate-envers', version: '5.4.27.Final'
         compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.11'
-        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-        implementation 'org.springframework.boot:spring-boot-starter-hateoas'
-        implementation 'org.springframework.boot:spring-boot-starter-web'
-        implementation 'org.springframework.boot:spring-boot-starter'
+
+        implementation group: 'javax.validation', name: 'validation-api', version: '2.0.1.Final'
         testCompile('org.springframework.boot:spring-boot-starter-test')
         compileOnly 'org.projectlombok:lombok'
         runtimeOnly 'mysql:mysql-connector-java'
         annotationProcessor 'org.projectlombok:lombok'
+
 
         testImplementation 'org.junit.jupiter:junit-jupiter-api:5.4.2'
         testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.4.2'
