+++ controller/build.gradle
@@ -4,10 +4,6 @@ plugins {
 
 dependencies {
     implementation project(':service')
-
-
-
-
 }
 
 bootJar{
