+++ controller/src/main/java/com/epam/esm/controller/config/ControllerConfig.java
@@ -25,8 +25,4 @@ public class ControllerConfig {
         return source;
     }
 
-
-
-
-
 }
