+++ controller/src/main/java/com/epam/esm/controller/controllers/OrderController.java
@@ -4,7 +4,7 @@ import com.epam.esm.controller.util.HateoasBuilder;
 import com.epam.esm.controller.util.PaginationBuilder;
 import com.epam.esm.entity.OrderCreationParameter;
 import com.epam.esm.entity.OrderDto;
-import com.epam.esm.service.impl.OrderService;
+import com.epam.esm.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.http.HttpStatus;
