+++ repository/src/main/java/com/epam/esm/entity/SearchGiftCertificateParameter.java
@@ -16,11 +16,11 @@ import java.util.List;
 public class SearchGiftCertificateParameter {
     private String name;
     private String description;
-    private List<String> tagNames;
+    private List<String> tags;
     private SortParameter sortBy;
     private OrderType order;
 
     public boolean isEmpty() {
-        return (name == null || name.isEmpty()) && (description == null || description.isEmpty()) && (tagNames == null || tagNames.isEmpty()) && sortBy == null && order == null;
+        return (name == null || name.isEmpty()) && (description == null || description.isEmpty()) && (tags == null || tags.isEmpty()) && sortBy == null && order == null;
     }
 }
