+++ repository/src/main/java/com/epam/esm/utils/GiftCertificateCriteriaBuilder.java
@@ -19,11 +19,10 @@ public class GiftCertificateCriteriaBuilder {
     public static final String REGEX = "%";
     public static final String PARAMETER_NAME = "name";
     public static final String PARAMETER_DESCRIPTION = "description";
-    public static final String PARAMETER_TAG_NAMES = "tagNames";
+    public static final String PARAMETER_TAG_NAMES = "tags";
     public static final String PARAMETER_SORTBY_NAME = "name";
     public static final String PARAMETER_SORTBY_CREATE_DATE = "createDate";
-    public static final String PARAMETER_ORDER = "order";
-
+    public static final String TAG_LIST_IN_GIFTCERTIFICATE = "tags";
     public static final String QUERY_SELECT_BY_TAG_NAME = "from Tag tag where tag.nameTag in (:" + PARAMETER_TAG_NAMES + ")";
 
 
@@ -43,20 +42,20 @@ public class GiftCertificateCriteriaBuilder {
 
         String parameterName = parameter.getName();
 
-        if (parameterName != null) {
-            Predicate predicate = criteriaBuilder.like(root.get(PARAMETER_NAME), REGEX
-                    + PARAMETER_NAME + REGEX);
+        if (Objects.nonNull(parameterName)) {
+            String pattern = REGEX + parameterName + REGEX;
+            Predicate predicate = criteriaBuilder.like(root.get(PARAMETER_NAME), pattern);
             predicateList.add(predicate);
         }
 
         String parameterDescription = parameter.getDescription();
-        if (parameterDescription != null) {
-            Predicate predicate = criteriaBuilder.like(root.get(PARAMETER_DESCRIPTION), REGEX
-                    + PARAMETER_DESCRIPTION + REGEX);
+        if (Objects.nonNull(parameterDescription)) {
+            String pattern = REGEX + parameterDescription + REGEX;
+            Predicate predicate = criteriaBuilder.like(root.get(PARAMETER_DESCRIPTION), pattern);
             predicateList.add(predicate);
         }
 
-        List<String> tagNames = parameter.getTagNames();
+        List<String> tagNames = parameter.getTags();
 
         if (Objects.nonNull(tagNames)) {
             List<String> distinctTagList = tagNames.stream().distinct().collect(Collectors.toList());
@@ -64,16 +63,16 @@ public class GiftCertificateCriteriaBuilder {
             if (tags.size() != distinctTagList.size()) {
                 return giftCertificateCriteriaQuery;
             }
-            tags.forEach(tag -> predicateList.add(criteriaBuilder.isMember(tag, root.get("tags"))));
+            tags.forEach(tag -> predicateList.add(criteriaBuilder.isMember(tag, root.get(TAG_LIST_IN_GIFTCERTIFICATE))));
 
         }
         giftCertificateCriteriaQuery.select(root).where(predicateList.toArray(new Predicate[0]));
 
         SortParameter sortParameter = parameter.getSortBy();
 
-        String sortValue = null;
+        String sortValue=null;
 
-        if (sortParameter != null) {
+        if (Objects.nonNull(sortParameter)) {
             switch (sortParameter) {
                 case NAME:
                     sortValue = PARAMETER_SORTBY_NAME;
@@ -86,7 +85,7 @@ public class GiftCertificateCriteriaBuilder {
 
         OrderType orderType = parameter.getOrder();
         Order order = null;
-        if (orderType != null) {
+        if (Objects.nonNull(orderType)) {
             switch (orderType) {
                 case ASC:
                     order = criteriaBuilder.asc(root.get(sortValue));
@@ -98,7 +97,6 @@ public class GiftCertificateCriteriaBuilder {
 
             giftCertificateCriteriaQuery.orderBy(order);
         }
-
 
         return giftCertificateCriteriaQuery;
