+++ repository/src/main/resources/application.properties
@@ -1,6 +1,11 @@
-spring.datasource.url = jdbc:mysql://127.0.0.1:3306/module02?serverTimezone=UTC&useSSL=false&allowPublicKeyRetrieval=true
-spring.datasource.username = root
-spring.datasource.password = 24081984
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/module02?serverTimezone=UTC&useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.data-username=root
+spring.datasource.data-password=24081984
+
+
+management.endpoint.beans.enabled=true
+management.endpoints.web.exposure.include=*
+
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=none
