+++ service/build.gradle
@@ -3,9 +3,6 @@ plugins {
 }
 
 dependencies {
-    implementation 'org.junit.jupiter:junit-jupiter:5.4.2'
-    implementation group: 'javax.validation', name: 'validation-api', version: '2.0.1.Final'
-    implementation 'org.springframework.boot:spring-boot-starter'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.4.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.4.2'
