+++ service/src/main/java/com/epam/esm/service/TagService.java
@@ -1,4 +1,4 @@
-package com.epam.esm.service.impl;
+package com.epam.esm.service;
 
 import com.epam.esm.entity.Tag;
 import com.epam.esm.entity.TagDto;
