+++ service/src/test/java/com/epam/esm/service/impl/GiftCertificateServiceTest.java
@@ -5,6 +5,7 @@ import com.epam.esm.entity.GiftCertificateDto;
 import com.epam.esm.entity.Tag;
 import com.epam.esm.entity.TagDto;
 import com.epam.esm.repository.GiftCertificateRepository;
+import com.epam.esm.service.GiftCertificateService;
 import com.epam.esm.service.exception.GiftCertificateDtoValidationException;
 import com.epam.esm.service.exception.NoSuchResourceException;
 import com.epam.esm.service.mapper.CertificateDtoMapper;
